+++ doc/changelog.rst
@@ -1,6 +1,7 @@
 Changelog
 =========
 
+* :bug:`606 major` Swallow python warnings during ast.parse
 * :feature:`-` Slash now detects test functions being redefined, hiding previous tests, and warns about it
 * :feature:`556` Long variable representations are now capped by default when distilling tracebacks
 * :feature:`-` Assertions coming from plugins and modules loaded from the project's ``.slashrc`` now also have assertion rewriting introspection enabled